http: only provide "preauth" for certain mechs
When a user specifies a username and password before being prompted
(through the callback), we similarly want to provide those to the remote
server immediately, without it returning a 401.  This saves a
round-trip.

However, only certain mechanisms support this behavior - for example,
Basic.  As we add other mechanisms that support username/password
credentials, we want to ensure that we do not try to preauthenticate
with them, since they require negotiation.  If we are given a username
and password, force the use of Basic.