http: match mechs with the cred types they support
Record the server's mechanisms, along with the supported credential
types.  We will prompt for credentials with the callback, then we must
ensure that we do not assume that a server can support any mech for the
given type.  Instead, use the first mechanism that the server supports
that also supports the credential type we want to use.