http: some auth types can preauthenticate
Teach the HTTP transport that not all credentials can be provided to the
server when they're supplied as "preauthentication" credentials, like in
a URL.  This works for basic, but it doesn't work for many other
credential types, like NTLM.

Also, ensure that the server supports the particular credential type
that we're specifying.  Previously we considered credential types as an
input to an auth mechanism - since the HTTP transport only supported
default credentials (via negotiate) and username/password credentials
(via basic), this worked.  However, if we are to add another mechanism
that uses username/password credentials, we'll need to be careful to
identify the types that are accepted.