proxy: handle proxy authentication replays
Some HTTP authentication mechanisms may require multiple round-trips
(like NTLM).  We need to handle authentication replays in the proxy
connection mechanism; we need to reconnect fully when we are not in
a keep-alive connection to the server; when we are in a keep-alive
scenario, we should instead resend a connection.