checkout: filter pathspecs for _all_ checkout types
We were previously applying the pathspec filter for the baseline
iterator during checkout, as well as the target tree.  This was an
oversight; in fact, we should apply the pathspec filter to _all_
checkout targets, not just trees.

Add a helper function to set the iterator pathspecs from the given
checkout pathspecs, and call it everywhere.